orqon_core/requirements.txt
# Core Dependencies (as per directive)
fastapi>=0.109.0
uvicorn>=0.27.0
pandas>=2.2.0
openpyxl>=3.1.2
chromadb>=0.4.22
pydantic>=2.6.0
streamlit>=1.28.0
requests>=2.31.0

# IBM watsonx Orchestrate (MCP toolkit)
ibm-watsonx-orchestrate>=1.15.0
ibm-watsonx-ai>=1.0.0

# Additional Required Packages
sentence-transformers>=2.3.1
python-dotenv>=1.0.1
langchain>=0.1.0
langchain-groq>=0.0.1
langgraph>=0.0.30
SpeechRecognition>=3.10.0
sounddevice>=0.5.0
soundfile>=0.13.0
numpy>=2.0.0
simple-salesforce>=1.12.0

# Google Cloud APIs
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.117.0

# Chart Generation
matplotlib>=3.8.0

# Document Parsing (IBM Document Extractor compatible)
PyPDF2>=3.0.0
python-docx>=1.1.0
Pillow>=10.0.0
pytesseract>=0.3.10
python-multipart>=0.0.6

# IBM Watson Speech Services
ibm-watson>=8.0.0
SpeechRecognition>=3.10.0
pydub>=0.25.1